[Input] Added option to set the keyboard buffer address (default use BIOS buffers). [Build tool] Added support for ROM segment in sub-directory using ProjSegments option. [Build tool] Added hotfix for SDCC 4.3.0 path error for CC1. [Build tool] Added support for Emulicious option to activate SCC. [Build tool] Adding new target format to emulate 16 KB ROM mapper using MSX-DOS2 RAM mapper (WIP). [Sample] Moved some sample segment to directory /segment/. Fixed several compile warning. Fixed SDCC 4.3.0 issue with '#' inside macro.

+++ projects/samples/s_ayfx.c
@@ -417,7 +417,7 @@ void main()
 
 	Print_SetPosition(0, PSG_Y);
 	Print_DrawText(" PSG\n\n");		
-	Draw_LineH(0, 255, PSG_Y+10, 0x5, 0);
+	Draw_LineH(0, 255, (u8)(PSG_Y + 10), 0x5, 0);
 	Print_DrawText("Chan. A\n");		
 	Print_DrawText("Chan. B\n");
 	Print_DrawText("Chan. C\n");
